v0.1.3
Byteferret now officially released.

JaroWinkler and Byteferret are now streamlined without string
compaction.

Improvement in sanitizeWord, it is now doesn't return string but
manipulating the string directly.

Improvement in uniqSentences, it is now doesn't return slice of slice of
strings, but manipulating it directly.

uniqSentences doesn't just filter duplicate sentences, it is now do
proper CSIS with extra JaroWinkler before that. Which contributes to the
slower benchmark results but more correct implementation of LexRank.

Adding new benchmark in README.md